+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.0.1] - 2024-01-31
+
+### Added
+- Payment Information in Order View (Backend)
+- Checkout Billing Address in Payment
+- Credentials Verification (Backend)
+
+### Updated
+- Webhook Web API
+- LESS Optimization
+- Template improvements
+
 ## [1.0.0] - 2023-12-18
 
 ### Added